Add RentController

+++ src/main/java/com/sda/dmpcars/service/RentService.java
@@ -33,7 +33,6 @@ public class RentService {
             Rent rent = rentDao.save(convertToRent(rentDto));
             return convertToRentDto(rent);
         }
-
         return new RentDto();
     }
 
@@ -68,14 +67,17 @@ public class RentService {
     public RentDto updateRent(RentDto rentDto) {
         Rent rent = convertToRent(rentDto);
         rent.setId(rentDto.getId());
-
         return convertToRentDto(rentDao.save(rent));
     }
 
-    public void deleteRent(RentDto rentDto) {
+    public boolean deleteRentById(RentDto rentDto) {
+        if (!rentDao.existsById(rentDto.getId())) {
+            return false;
+        }
         Rent rent = convertToRent(rentDto);
         rent.setId(rentDto.getId());
         rentDao.delete(rent);
+        return true;
     }
 
     public void deleteAllRents() {
@@ -90,9 +92,15 @@ public class RentService {
      * @return RentDao converted to Rent
      */
     private Rent convertToRent(RentDto rentDto) {
-        return Rent.builder().fromDate(rentDto.getFromDate()).toDate(rentDto.getToDate())
+        return Rent
+                .builder()
+                .fromDate(rentDto.getFromDate())
+                .toDate(rentDto.getToDate())
                 .totalPrice(rentDto.getTotalPrice())
-                .account(Account.builder().id(rentDto.getAccountDto().getId()).username(rentDto.getAccountDto().getUsername())
+                .account(Account
+                        .builder()
+                        .id(rentDto.getAccountDto().getId())
+                        .username(rentDto.getAccountDto().getUsername())
                         .build())
                 .car(Car.builder().id(rentDto.getCarDto().getId()).build())
                 .build();
